+++ src/components/CategoryBanner.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { Icon } from '@material-ui/core';
-import { ArrowBack } from '@material-ui/icons';
+
 import { colors, font } from '../App.styles';
+import BackButton from './BackButton';
 
 interface Props {
   color?: string;
@@ -19,28 +18,6 @@ const CategoryBannerContainer = styled.div`
   wrap: no-wrap;
 `;
 
-const CategoryBannerLink = styled(Link)`
-  align-items: center;
-  display: flex;
-  padding-right: ${font.helpers.convertPixelsToRems(12)};
-`;
-
-const CategoryBannerIcon = styled(Icon)`
-  && {
-    align-items: center;
-    display: flex;
-    font-size: ${font.helpers.convertPixelsToRems(36)};
-    height: auto;
-    width: auto;
-  }
-` as typeof Icon;
-
-const CategoryBannerArrowBack = styled(ArrowBack)`
-  && {
-    font-size: inherit;
-  }
-` as typeof ArrowBack;
-
 const CategoryBannerHeading = styled.h1`
   align-items: center;
   display: flex;
@@ -51,11 +28,7 @@ const CategoryBannerHeading = styled.h1`
 
 const CategoryBanner = ({ color, text }: Props) => (
   <CategoryBannerContainer color={color}>
-    <CategoryBannerLink to="/">
-      <CategoryBannerIcon>
-        <CategoryBannerArrowBack />
-      </CategoryBannerIcon>
-    </CategoryBannerLink>
+    <BackButton />
     <CategoryBannerHeading>{text}</CategoryBannerHeading>
   </CategoryBannerContainer>
 );
